Make publish script work remotely
Instead of copying content from one local directory to another,
publish.sh will not use ssh to actually delete and copy the files
directly on graemehill.ca. This means that there is no need to manually
SSH to the destination server or to have a working copy of the git repo
on the web server.